Code ref Support for .NET 8 Drop support for .NET 6
