+++ DataAnnotatedModelValidations.Tests/.config/dotnet-tools.json
@@ -3,7 +3,7 @@
   "isRoot": true,
   "tools": {
     "dotnet-reportgenerator-globaltool": {
-      "version": "5.1.24",
+      "version": "5.2.0",
       "commands": [
         "reportgenerator"
       ]
