+++ DataAnnotatedModelValidations.Tests/Extensions/CollectionExtensionsTests.cs
@@ -1,6 +1,9 @@
+using System.Linq;
+using DataAnnotatedModelValidations.Extensions;
+
 namespace DataAnnotatedModelValidations.Tests;
 
-public class ExtensionsTests
+public class CollectionExtensionsTests
 {
     [Fact(DisplayName = "AsEnumerable - Enumerate Value")]
     public void EnumerateValue() => Assert.Contains("test", "test".AsEnumerable());
