+++ DataAnnotatedModelValidations/Attributes/IgnoreModelValidationAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace DataAnnotatedModelValidations;
+namespace DataAnnotatedModelValidations.Attributes;
 
 [AttributeUsage(AttributeTargets.Parameter)]
 public class IgnoreModelValidationAttribute : Attribute { }
