+++ DataAnnotatedModelValidations/DataAnnotatedModelValidations.csproj
@@ -1,13 +1,12 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net6.0;net7.0</TargetFrameworks>
-    <LangVersion>11.0</LangVersion>
+    <TargetFrameworks>net7.0;net8.0</TargetFrameworks>
     <Nullable>enable</Nullable>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
     <PackageId>$(MSBuildProjectName)</PackageId>
-    <Version>4.2.0</Version>
+    <Version>5.0.0</Version>
     <Authors>Adamos Fiakkas</Authors>
     <Description>Data Annotated Model Validation Middleware for HotChocolate</Description>
     <Copyright>Adamos Fiakkas</Copyright>
@@ -18,7 +17,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="HotChocolate.Execution" Version="13.6.0" />
+    <PackageReference Include="HotChocolate.Execution" Version="13.7.0" />
     <PackageReference Include="Humanizer" Version="2.14.1" />
   </ItemGroup>
