+++ DataAnnotatedModelValidations/DataAnnotationsValidatorRegistration.cs
@@ -1,9 +1,11 @@
-﻿using HotChocolate.Execution.Configuration;
+﻿using DataAnnotatedModelValidations.Middleware;
+using DataAnnotatedModelValidations.TypeInterceptors;
+using HotChocolate.Execution.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace DataAnnotatedModelValidations;
 
-public static class RegisterValidatorMiddleware
+public static class DataAnnotationsValidatorRegistration
 {
     public static IRequestExecutorBuilder AddDataAnnotationsValidator(this IRequestExecutorBuilder requestExecutorBuilder) =>
         requestExecutorBuilder
